+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.16] - 2024-05-15
+
+### Changed
+- [matter.js]: Update to @project-chip/matter-node.js v. 0.9.0
+
+### Added
+
+- [frontend]: Frontend updated to 1.0.0.
+- [frontend]: Added check version interval.
+- [frontend]: Added help and version information icons for plugins.
+- [frontend]: Added version information for Matterbridge (Click on the version badge).
+- [frontend]: Added help badge for Matterbridge.
+- [frontend]: Added version information badge for Matterbridge.
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="./yellow-button.png" alt="Buy me a coffee" width="120">
+</a>
+
 ## [1.2.15] - 2024-05-14
 
 ### Added